split project into modules
