+++ common/src/main/java/cf/baradist/Edge.java
