+++ dijkstra/build.gradle
@@ -1,0 +1,3 @@
+shadowJar {
+    baseName = 'dijkstra'
+}
