+++ dijkstra/src/main/java/cf/baradist/Dijkstra.java
@@ -1,35 +1,12 @@
 package cf.baradist;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.Stack;
-import java.util.stream.Collectors;
 
-import static java.lang.Double.NaN;
-
-public class Dijkstra {
+public class Dijkstra extends AbstractPathFinder {
     private Set<Integer> visitedIndexes;
-    private Double[][] matrix;
-    private int size;
     private Double[] d;
-    private int from;
-
-    public void readGraphMatrix(String filename) {
-        try {
-            matrix = Files.lines(Paths.get(filename))
-                    .map(line ->
-                            Arrays.stream(line.split("\\s+"))
-                                    .map(Dijkstra::getDoubleValueOf).toArray(Double[]::new))
-                    .toArray(Double[][]::new);
-            size = matrix.length;
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
 
     public void findWeights(int from, int to) {
         this.from = from;
@@ -49,7 +26,18 @@ public class Dijkstra {
             current = nextTheClosestIndex();
             visitedIndexes.add(current);
         }
+
         System.out.println("The distance from " + from + " to " + to + " is: " + d[to]);
+    }
+
+    private void updateWeightsToNeighbors(int current, Double[] currentLine) {
+        for (int i = 0; i < size; i++) {
+            if (!isRealValue(currentLine[i]) || i == current
+                    || visitedIndexes.contains(i)) {
+                continue;
+            }
+            d[i] = Math.min(d[i], d[current] + currentLine[i]);
+        }
     }
 
     public Stack<Edge> findPath(int to) {
@@ -64,24 +52,6 @@ public class Dijkstra {
         return path;
     }
 
-    public int getSize() {
-        return size;
-    }
-
-    public void setMatrix(Double[][] matrix) {
-        size = matrix.length;
-        this.matrix = matrix;
-    }
-
-    private static Double getDoubleValueOf(String string) {
-        if ("-".equals(string)) {
-            return NaN;
-        } else if ("∞".equals(string)) {
-            return Double.POSITIVE_INFINITY;
-        }
-        return Double.valueOf(string);
-    }
-
     private int findPrev(int current) {
         for (int i = 0; i < size; i++) {
             if (isRealValue(matrix[i][current]) && visitedIndexes.contains(i)) {
@@ -91,20 +61,6 @@ public class Dijkstra {
             }
         }
         throw new IllegalArgumentException("Can't find a path");
-    }
-
-    private void updateWeightsToNeighbors(int current, Double[] currentLine) {
-        for (int i = 0; i < size; i++) {
-            if (!isRealValue(currentLine[i]) || i == current
-                    || visitedIndexes.contains(i)) {
-                continue;
-            }
-            d[i] = Math.min(d[i], d[current] + currentLine[i]);
-        }
-    }
-
-    private static boolean isRealValue(Double value) {
-        return !(value.isNaN() || value.isInfinite());
     }
 
     private int nextTheClosestIndex() {
@@ -129,15 +85,5 @@ public class Dijkstra {
             }
         }
         return minIndex;
-    }
-
-    private String toString(Double[] doubles) {
-        return Arrays.stream(doubles)
-                .map(Dijkstra::toString)
-                .collect(Collectors.joining("\t"));
-    }
-
-    private static String toString(Double value) {
-        return value == null ? "null" : value.isInfinite() ? "∞" : String.valueOf(value);
     }
 }
