+++ dijkstra/src/main/java/cf/baradist/Main.java
