+++ dijkstra/src/test/java/cf/baradist/DijkstraTest.java
