+++ dijkstra/src/test/resources/dijkstra_18.txt
@@ -1,4 +1,4 @@
--   11  14  ∞   ∞   ∞
+- 	11	14	∞	∞	∞
 ∞   -   8   10  15  ∞
 ∞   ∞   -   11  16  20
 ∞   ∞   ∞   -   ∞   12
